+++ src/hex_grid.rs
@@ -1,28 +1,24 @@
 use numpy::ndarray::*;
 use crate::utils::*;
 
-fn iseven(val: i64) -> bool {
-    val % 2 == 0
-}
-
-fn modulus(val: f64, modulus: f64) -> f64 {
-    ((val % modulus) + modulus ) % modulus
-}
 pub struct HexGrid {
     pub cellsize: f64,
     pub offset: (f64, f64),
+    pub rotation: f64,
+    pub rotation_matrix: Array2<f64>,
+    pub rotation_matrix_inv: Array2<f64>,
 }
 
 impl HexGrid {
-    pub fn new(cellsize: f64, offset: (f64, f64)) -> Self {
-        // Note: dx and dy are required to normalize the offset
-        //       Create an intermediate grid with incorrect offset first,
-        //       such that dx and dy of the intermediate grid can be used
-        //       to calculate dx and dy. Then create self with the
-        //       normalized offset
-        let self_tmp = HexGrid { cellsize, offset };
+    pub fn new(cellsize: f64, offset: (f64, f64), rotation: f64) -> Self {
+        let rotation_matrix = _rotation_matrix(rotation);
+        let rotation_matrix_inv = _rotation_matrix(-rotation);
+        // TODO: Find a way to normalize_offset without having to instantiate tmp object
+        let self_tmp = HexGrid { cellsize, offset, rotation, rotation_matrix, rotation_matrix_inv };
         let offset = normalize_offset(offset, self_tmp.dx(), self_tmp.dy());
-        HexGrid { cellsize, offset }
+        let rotation_matrix = _rotation_matrix(rotation);
+        let rotation_matrix_inv = _rotation_matrix(-rotation);
+        HexGrid { cellsize, offset, rotation, rotation_matrix, rotation_matrix_inv }
     }
 
     pub fn radius(&self) -> f64 {
@@ -41,14 +37,22 @@ impl HexGrid {
         let mut centroids = Array2::<f64>::zeros((index.shape()[0], 2));
 
         for cell_id in 0..centroids.shape()[0] {
-            let point = self.centroid_single_point(index[Ix2(cell_id, 0)], index[Ix2(cell_id, 1)]);
+            let point = self.centroid_xy_no_rot(index[Ix2(cell_id, 0)], index[Ix2(cell_id, 1)]);
             centroids[Ix2(cell_id, 0)] = point.0;
             centroids[Ix2(cell_id, 1)] = point.1;
+        }
+
+        if self.rotation != 0. {
+            for cell_id in 0..centroids.shape()[0] {
+                let mut centroid = centroids.slice_mut(s![cell_id, ..]);
+                let cent_rot = self.rotation_matrix.dot(&centroid);
+                centroid.assign(&cent_rot);
+            }
         }
         centroids
     }
 
-    fn centroid_single_point(&self, x: i64, y: i64) -> (f64, f64) {
+    fn centroid_xy_no_rot(&self, x: i64, y: i64) -> (f64, f64) {
         let mut centroid_x = x as f64 * self.dx() + (self.dx() / 2.) + self.offset.0;
         let centroid_y = y as f64 * self.dy() + (self.dy() / 2.) + self.offset.1;
 
@@ -61,8 +65,10 @@ impl HexGrid {
     pub fn cell_at_location(&self, points: &ArrayView2<f64>) -> Array2<i64> {
         let mut index = Array2::<i64>::zeros((points.shape()[0], 2));
         for cell_id in 0..points.shape()[0] {
-            let x = points[Ix2(cell_id, 0)];
-            let y = points[Ix2(cell_id, 1)];
+            let point = points.slice(s![cell_id, ..]);
+            let point = self.rotation_matrix_inv.dot(&point);
+            let x = point[Ix1(0)];
+            let y = point[Ix1(1)];
 
             // determine initial id_y
             let mut id_y = ((y - self.radius() / 4.) / self.dy()).floor();
@@ -124,9 +130,19 @@ impl HexGrid {
             for corner_id in 0..6 {
                 let angle_deg = 60. * corner_id as f64 - 30.;
                 let angle_rad = angle_deg * std::f64::consts::PI / 180.;
-                let centroid = self.centroid_single_point(index[Ix2(cell_id, 0)], index[Ix2(cell_id, 1)]);
+                let centroid = self.centroid_xy_no_rot(index[Ix2(cell_id, 0)], index[Ix2(cell_id, 1)]);
                 corners[Ix3(cell_id, corner_id, 0)] = centroid.0 + self.radius() * angle_rad.cos();
                 corners[Ix3(cell_id, corner_id, 1)] = centroid.1 + self.radius() * angle_rad.sin();
+            }
+        }
+
+        if self.rotation != 0. {
+            for cell_id in 0..corners.shape()[0] {
+                for corner_id in 0..corners.shape()[1] {
+                    let mut corner_xy = corners.slice_mut(s![cell_id, corner_id, ..]);
+                    let rotated_corner_xy = self.rotation_matrix.dot(&corner_xy);
+                    corner_xy.assign(&rotated_corner_xy);
+                }
             }
         }
         corners
@@ -142,9 +158,21 @@ impl HexGrid {
         let mut nearby_cells = Array3::<i64>::zeros((points.shape()[0], 3, 2));
 
         let cell_ids = self.cell_at_location(points);
+
+        // FIXME: Find a way to not clone points in the case of no rotation
+        //        If points is made mutable within the conditional, it is dropped from scope and nothing changed
+        let mut points = points.to_owned();
+        if self.rotation != 0. {
+            for cell_id in 0..points.shape()[0] {
+                let mut point = points.slice_mut(s![cell_id, ..]);
+                let point_rot = self.rotation_matrix_inv.dot(&point);
+                point.assign(&point_rot);
+            }
+        }
+
         for cell_id in 0..points.shape()[0] {
             // Determine the azimuth based on the direction vector from the cell centroid to the point
-            let centroid = self.centroid_single_point(cell_ids[Ix2(cell_id, 0)], cell_ids[Ix2(cell_id, 1)]);
+            let centroid = self.centroid_xy_no_rot(cell_ids[Ix2(cell_id, 0)], cell_ids[Ix2(cell_id, 1)]);
             let direction_x = points[Ix2(cell_id, 0)] - centroid.0;
             let direction_y = points[Ix2(cell_id, 1)] - centroid.1;
             let mut azimuth = direction_x.atan2(direction_y) * 180. / std::f64::consts::PI;